vacation-planner/backend-python/requirements.txt
fastapi
uvicorn
pydantic>=2.0
pydantic-ai
groq
ddgs
python-dotenv
loguru
supabase
python-jose[cryptography]
